Changed File names and created a splits tab to add workouts to a weekly split plan
