--- app/(tabs)/gymfinder.tsx
@@ -1,10 +1,0 @@
-import { View, Text } from 'react-native';
-
-export default function gymfinder() {
-    return(
-    <View>
-        <Text>
-            Welcome to the gymfinder Page
-        </Text>
-    </View>)
-}
