+++ app/(tabs)/splits.tsx
@@ -1,206 +1,114 @@
-import React, { useState, useEffect } from 'react';
-import { View, Text, FlatList, TextInput, Button, StyleSheet, ScrollView, TouchableOpacity } from 'react-native';
-import axios from 'axios';
-import { Exercise } from '../models/exerciseModel';
+import React, { useState, useEffect } from "react";
+import { View, Text, TextInput, Button, FlatList, ScrollView } from "react-native";
+import BouncyCheckbox from "react-native-bouncy-checkbox";
+import { db } from "../firebaseConfig"; // Import Firestore from your setup
+import { doc, getDoc, setDoc } from "firebase/firestore";
+import { auth } from "../firebaseConfig"; // Import Firebase Auth to get user ID
 
-const Index = () => {
-  const [exercises, setExercises] = useState<Exercise[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [searchTerm, setSearchTerm] = useState<string>(''); // State for search term
-  const [expandedInstructions, setExpandedInstructions] = useState<Set<number>>(new Set()); // Track expanded instructions
+const WorkoutPlanScreen = () => {
+  const userID = auth.currentUser?.uid;
+  const [workoutPlan, setWorkoutPlan] = useState<{ [key: string]: { split: string; workouts: { name: string; completed: boolean }[] } }>({});
+  const [newWorkout, setNewWorkout] = useState<{ [key: string]: string }>({});
 
-  const fetchExercises = async (query: string) => {
-    try {
-      setLoading(true); // Show loading spinner
-      const response = await axios.get('http://localhost:5000/', {
-        params: {
-          name: query, // Pass search term to backend
-        },
-      });
-      setExercises(response.data.exercises); // Set exercises from API response
-    } catch (error) {
-      setError('Failed to fetch exercises');
-      console.error(error);
-    } finally {
-      setLoading(false); // Hide loading spinner
-    }
-  };
-
-  const handleSearch = () => {
-    if (searchTerm.trim() !== '') {
-      fetchExercises(searchTerm); // Fetch exercises when user clicks search button
-    }
-  };
-
-  const handleEnterPress = (e: any) => {
-    if (e.key === 'Enter' && searchTerm.trim() !== '') {
-      fetchExercises(searchTerm); // Fetch exercises when user presses Enter
-    }
-  };
-
-  const toggleInstructions = (index: number) => {
-    setExpandedInstructions((prevState) => {
-      const newState = new Set(prevState);
-      if (newState.has(index)) {
-        newState.delete(index); // Collapse
-      } else {
-        newState.add(index); // Expand
+  // Load workout plan from Firestore
+  useEffect(() => {
+    const fetchWorkoutPlan = async () => {
+      if (!auth.currentUser) return;
+      const userRef = doc(db, "users", auth.currentUser.uid);
+      const docSnap = await getDoc(userRef);
+      if (docSnap.exists()) {
+        setWorkoutPlan(docSnap.data()?.workoutPlans || {});
       }
-      return newState;
+    };
+  
+    fetchWorkoutPlan();
+  }, []);
+
+  const saveWorkoutPlan = async () => {
+    console.log("Saving workout plan:", workoutPlan);
+    const user = auth.currentUser;
+    if (!user) {
+      console.error("No user logged in.");
+      return;
+    }
+
+    try {
+      const userRef = doc(db, "users", user.uid);
+      await setDoc(userRef, { workoutPlans: workoutPlan }, { merge: true });
+      console.log("Workout plan saved successfully!");
+    } catch (error) {
+      console.error("Error saving workout plan:", error);
+    }
+  };
+  
+
+  // Update split name
+  const updateSplit = (day: string, split: string) => {
+    setWorkoutPlan((prev) => ({
+      ...prev,
+      [day]: { ...prev[day], split, workouts: prev[day]?.workouts || [] },
+    }));
+  };
+
+  // Add new workout
+  const addWorkout = (day: string) => {
+    if (!newWorkout[day]) return;
+    setWorkoutPlan((prev) => ({
+      ...prev,
+      [day]: {
+        ...prev[day],
+        workouts: [...(prev[day]?.workouts || []), { name: newWorkout[day], completed: false }],
+      },
+    }));
+    setNewWorkout((prev) => ({ ...prev, [day]: "" })); // Clear input field
+  };
+
+  // Toggle workout completion
+  const toggleWorkoutCompletion = (day: string, index: number) => {
+    setWorkoutPlan((prev) => {
+      const updatedWorkouts = [...prev[day].workouts];
+      updatedWorkouts[index].completed = !updatedWorkouts[index].completed;
+      return { ...prev, [day]: { ...prev[day], workouts: updatedWorkouts } };
     });
   };
 
-  if (loading) {
-    return (
-      <View style={styles.center}>
-        <Text style={styles.loadingText}>Loading...</Text>
-      </View>
-    );
-  }
-
-  if (error) {
-    return (
-      <View style={styles.center}>
-        <Text style={styles.errorText}>{error}</Text>
-      </View>
-    );
-  }
-
   return (
-    <ScrollView style={styles.container}>
-      <Text style={styles.title}>Exercises:</Text>
-
-      {/* Search Bar */}
-      <TextInput
-        style={styles.searchBar}
-        placeholder="Search exercises..."
-        value={searchTerm}
-        onChangeText={setSearchTerm}
-        onSubmitEditing={handleEnterPress} // Trigger search when pressing "Enter"
-      />
-
-      {/* Search Button */}
-      <Button title="Search" onPress={handleSearch} />
-
-      {/* Display Exercises */}
-      {exercises.length > 0 && (
-        <FlatList
-          data={exercises}
-          keyExtractor={(item, index) => index.toString()}
-          renderItem={({ item, index }) => (
-            <View style={styles.exerciseContainer}>
-              <Text style={styles.exerciseName}>{item.name}</Text>
-              <Text style={styles.exerciseDetails}>Target Muscle: {item.muscle}</Text>
-              <Text style={styles.exerciseDetails}>Type: {item.type}</Text>
-              <Text style={styles.exerciseDetails}>Equipment: {item.equipment}</Text>
-              <Text style={styles.exerciseDetails}>Difficulty: {item.difficulty}</Text>
-
-              {/* Instructions: Show a truncated version initially, with a "..." button to expand */}
-              <View>
-                <Text
-                  style={styles.instructions}
-                  numberOfLines={expandedInstructions.has(index) ? undefined : 2}
-                >
-                  {item.instructions || 'No instructions available'}
-                </Text>
-                {item.instructions && (
-                  <TouchableOpacity onPress={() => toggleInstructions(index)}>
-                    <Text style={styles.expandText}>
-                      {expandedInstructions.has(index) ? 'Show Less' : '...'}
-                    </Text>
-                  </TouchableOpacity>
-                )}
+    <ScrollView contentContainerStyle={{ flexGrow: 1, padding: 20 }}>
+      {["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"].map((day) => (
+        <View key={day} style={{ marginBottom: 20 }}>
+          <Text style={{ fontSize: 18, fontWeight: "bold" }}>{day}</Text>
+          <TextInput
+            placeholder="Enter split (e.g., Push, Pull, Legs)"
+            value={workoutPlan[day]?.split || ""}
+            onChangeText={(text) => updateSplit(day, text)}
+            style={{ borderBottomWidth: 1, marginBottom: 10 }}
+          />
+          <FlatList
+            data={workoutPlan[day]?.workouts || []}
+            keyExtractor={(item, index) => `${item.name}-${index}`}
+            renderItem={({ item, index }) => (
+              <View style={{ flexDirection: "row", alignItems: "center", marginBottom: 5 }}>
+                <BouncyCheckbox
+                  isChecked={item.completed}
+                  text={item.name}
+                  onPress={() => toggleWorkoutCompletion(day, index)}
+                />
               </View>
-            </View>
-          )}
-        />
-      )}
-
-      {/* Show "No exercises found" message only if search term is not empty and no exercises are found */}
-      {exercises.length === 0 && searchTerm.trim() !== '' && (
-        <Text style={styles.noResultsText}>No exercises found for "{searchTerm}"</Text>
-      )}
+            )}
+          />
+          <TextInput
+            placeholder="Enter workout (e.g., Bench Press)"
+            value={newWorkout[day] || ""}
+            onChangeText={(text) => setNewWorkout((prev) => ({ ...prev, [day]: text }))}
+            style={{ borderBottomWidth: 1, marginBottom: 5 }}
+          />
+          <Button title="Add Workout" onPress={() => addWorkout(day)} />
+        </View>
+      ))}
+      <Button title="Save Plan" onPress={saveWorkoutPlan} />
     </ScrollView>
+
   );
 };
 
-// Styles for the layout
-const styles = StyleSheet.create({
-  container: {
-    padding: 20,
-    backgroundColor: '#f5f5f5',
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 20,
-    textAlign: 'center',
-  },
-  searchBar: {
-    height: 40,
-    borderColor: '#ccc',
-    borderWidth: 1,
-    borderRadius: 8,
-    marginBottom: 20,
-    paddingHorizontal: 10,
-    fontSize: 16,
-  },
-  exerciseContainer: {
-    backgroundColor: '#fff',
-    padding: 15,
-    marginBottom: 15,
-    borderRadius: 8,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 6,
-    elevation: 3,
-  },
-  exerciseName: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    color: '#333',
-  },
-  exerciseDetails: {
-    fontSize: 16,
-    color: '#555',
-    marginVertical: 2,
-  },
-  instructions: {
-    fontSize: 14,
-    color: '#777',
-    marginTop: 10,
-    fontStyle: 'italic',
-  },
-  expandText: {
-    fontSize: 14,
-    color: '#007BFF',
-    marginTop: 5,
-    fontStyle: 'italic',
-  },
-  loadingText: {
-    fontSize: 20,
-    textAlign: 'center',
-    marginTop: 20,
-  },
-  errorText: {
-    fontSize: 20,
-    color: 'red',
-    textAlign: 'center',
-    marginTop: 20,
-  },
-  noResultsText: {
-    fontSize: 18,
-    textAlign: 'center',
-    color: 'gray',
-    marginTop: 20,
-  },
-  center: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    flex: 1,
-  },
-});
-
-export default Index;
+export default WorkoutPlanScreen;
