+++ package.json
@@ -16,6 +16,9 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^14.0.2",
+    "@react-native-async-storage/async-storage": "^1.24.0",
+    "@react-native-firebase/app": "^21.7.1",
+    "@react-native-firebase/firestore": "^21.7.1",
     "@react-navigation/bottom-tabs": "^7.0.0",
     "@react-navigation/native": "^7.0.0",
     "@react-navigation/native-stack": "^7.2.0",
@@ -33,10 +36,11 @@
     "expo-system-ui": "~4.0.7",
     "expo-updates": "~0.26.12",
     "expo-web-browser": "~14.0.2",
-    "firebase": "^11.1.0",
+    "firebase": "^11.2.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-native": "0.76.6",
+    "react-native-bouncy-checkbox": "^4.1.2",
     "react-native-gesture-handler": "~2.20.2",
     "react-native-reanimated": "~3.16.1",
     "react-native-safe-area-context": "4.12.0",
